bugfixes: 3076371
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@4441 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ Gizmos/newclassdlg.cpp
@@ -358,7 +358,7 @@ void NewClassDlg::GetNamespacesList(wxArrayString& namespacesArray)
 		return;
 
 	int prevPos = 0;
-	unsigned int pos = textNamespaces.find (wxT("::"), prevPos);
+	size_t pos = textNamespaces.find (wxT("::"), prevPos);
 
 	while (pos != wxString::npos) {
 		wxString token = textNamespaces.Mid(prevPos, pos-prevPos);
